test：rebuild test folder
