+++ account/nuls-account/src/main/java/io/nuls/account/init/AccountBootstrap.java
@@ -29,7 +29,6 @@ public class AccountBootstrap {
             initCfg();
             //读取配置文件，数据存储根目录，初始化打开该目录下所有表连接并放入缓存
             RocksDBService.init(AccountParam.getInstance().getDataPath());
-            RocksDBService.init(AccountParam.getInstance().getDataPath());
             //springLite容器初始化
             SpringLiteContext.init("io.nuls.account", new ModularServiceMethodInterceptor());
             //启动时间同步线程
